# docs/openapi.yml
openapi: 3.1.0
info:
  title: "Maroa.ai API"
  version: 2.3.0
  description: "AI marketing automation SaaS — Inngest + Express backend. See CLAUDE.md for architecture."
servers:
  - url: https://maroa-api-production.up.railway.app
    description: Production
  - url: http://localhost:3000
    description: "Local dev"
...
paths:
...
  /webhook/content-generate:
    post:
      tags:
        - content
      summary: "Generate single piece of content"
      description: "Single concept → image (Higgsfield) → caption + score → save as draft. Mounted with cost-guard middleware. Per-business rate limit."
      responses:
        200:
          description: "Successful response"
        400:
          description: "Validation error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        402:
          description: "Payment Required — cost cap reached"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        429:
          description: "Rate limited"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        500:
          description: "Server error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
      requestBody:
        description: "Request body — hand-edit shape per route"
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
  /webhook/content-pieces-get:
    get:
      tags:
        - webhook
      summary: "GET /webhook/content-pieces-get"
      description: "Maroa automation webhook (GET /webhook/content-pieces-get). Typically requires x-webhook-secret and business ownership when business_id is present."
      responses:
        200:
          description: "Successful response"
        400:
          description: "Validation error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        402:
          description: "Payment Required — cost cap reached"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        429:
          description: "Rate limited"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        500:
          description: "Server error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /webhook/content-approve:
    post:
      tags:
        - webhook
      summary: "POST /webhook/content-approve"
      description: "Maroa automation webhook (POST /webhook/content-approve). Typically requires x-webhook-secret and business ownership when business_id is present."
      responses:
        200:
          description: "Successful response"
        400:
          description: "Validation error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        402:
          description: "Payment Required — cost cap reached"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        429:
          description: "Rate limited"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        500:
          description: "Server error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
      requestBody:
        description: "Request body — hand-edit shape per route"
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    WebhookSecret:
      type: apiKey
      in: header
      name: x-webhook-secret
    MetricsToken:
      type: apiKey
      in: header
      name: x-metrics-token
      description: "METRICS_SCRAPE_TOKEN or ORCHESTRATOR_SECRET; Bearer auth also accepted"
  schemas:
    ErrorEnvelope:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: "Machine-readable error code (e.g. VALIDATION_ERROR, RATE_LIMITED, COST_CAP_REACHED)"
            message:
              type: string
            details:
              type: object
              additionalProperties: true
              nullable: true
            timestamp:
              type: string
              format: date-time
            request_id:
              type: string
...